LayoutTests/fast/canvas/webgl/context-attributes-alpha-depth-stencil-antialias-expected.txt

       
Verify WebGLContextAttributes are working as specified, including alpha, depth, stencil, antialias, but not premultipliedAlpha

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Regression test for https://bugs.webkit.org/show_bug.cgi?id=33416 : Hook up WebGLContextAttributes to OpenGL context creation code
Testing alpha = true
PASS webGL = getWebGL('alphaOn', { alpha: true, depth: false, stencil: false, antialias: false }, [ 0, 0, 0, 0 ], 1, 0) is non-null.
PASS contextAttribs = webGL.getContextAttributes() is non-null.
PASS contextAttribs.alpha is true
PASS contextAttribs.depth is false
PASS contextAttribs.stencil is false
PASS contextAttribs.antialias is false
PASS contextAttribs.premultipliedAlpha is true
PASS pixel is correctColor
Testing alpha = false
PASS webGL = getWebGL('alphaOff', { alpha: false, depth: false, stencil: false, antialias: false }, [ 0, 0, 0, 0 ], 1, 0) is non-null.
PASS contextAttribs = webGL.getContextAttributes() is non-null.
PASS contextAttribs.alpha is false
PASS contextAttribs.depth is false
PASS contextAttribs.stencil is false
PASS contextAttribs.antialias is false
PASS contextAttribs.premultipliedAlpha is true
PASS pixel is correctColor
Testing depth = true
PASS webGL = getWebGL('depthOn', { stencil: false, antialias: false }, [ 0, 0, 0, 1 ], 1, 0) is non-null.
PASS contextAttribs = webGL.getContextAttributes() is non-null.
PASS contextAttribs.depth is true
PASS contextAttribs.alpha is true
PASS contextAttribs.stencil is false
PASS contextAttribs.antialias is false
PASS contextAttribs.premultipliedAlpha is true
PASS pixel is correctColor
Testing depth = false
PASS webGL = getWebGL('depthOff', { depth: false, stencil: false, antialias: false }, [ 0, 0, 0, 1 ], 1, 0) is non-null.
PASS contextAttribs = webGL.getContextAttributes() is non-null.
PASS contextAttribs.depth is false
PASS contextAttribs.alpha is true
PASS contextAttribs.stencil is false
PASS contextAttribs.antialias is false
PASS contextAttribs.premultipliedAlpha is true
PASS pixel is correctColor
Testing stencil = true
PASS webGL = getWebGL('stencilOn', { depth: false, stencil: true, antialias: false }, [ 0, 0, 0, 1 ], 1, 0) is non-null.
PASS contextAttribs = webGL.getContextAttributes() is non-null.
PASS contextAttribs.depth == contextAttribs.stencil is true
PASS contextAttribs.alpha is true
PASS contextAttribs.antialias is false
PASS contextAttribs.premultipliedAlpha is true
PASS pixel is correctColor
Testing stencil = false
PASS webGL = getWebGL('stencilOff', { depth: false, stencil: false, antialias: false }, [ 0, 0, 0, 1 ], 1, 0) is non-null.
PASS contextAttribs = webGL.getContextAttributes() is non-null.
PASS contextAttribs.depth == contextAttribs.stencil is true
PASS contextAttribs.alpha is true
PASS contextAttribs.antialias is false
PASS contextAttribs.premultipliedAlpha is true
PASS pixel is correctColor
Testing antialias = true
PASS webGL = getWebGL('antialiasOn', { depth: false, stencil: false, alpha: false, antialias: true }, [ 0, 0, 0, 1 ], 1, 0) is non-null.
PASS contextAttribs = webGL.getContextAttributes() is non-null.
PASS contextAttribs.depth is false
PASS contextAttribs.stencil is false
PASS contextAttribs.alpha is false
PASS contextAttribs.antialias == true || contextAttribs.antialias == false is true
PASS contextAttribs.premultipliedAlpha is true
PASS pixel[0] == 255 || pixel[0] == 0 is false
Testing antialias = false
PASS webGL = getWebGL('antialiasOff', { depth: false, stencil: false, alpha: false, antialias: false }, [ 0, 0, 0, 1 ], 1, 0) is non-null.
PASS contextAttribs = webGL.getContextAttributes() is non-null.
PASS contextAttribs.depth is false
PASS contextAttribs.stencil is false
PASS contextAttribs.alpha is false
PASS contextAttribs.antialias == true || contextAttribs.antialias == false is true
PASS contextAttribs.premultipliedAlpha is true
PASS pixel[0] == 255 || pixel[0] == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
